+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Update base image to Ubuntu 16.04
 * Set version number in command prompt dynamically
 * Bluemix Volume support added. (See: `setup-var-mqm.sh`)
+* Added MQ Developer Defaults
 
 ## 1.0.2 (2016-11-02)
 * Add MQ V9